app/requirements.txt
cloudinary==1.10.0
marshmallow==3.0.0b5
motor==1.2.0
pymongo==3.6.0
python-consul
python-dateutil==2.6.1
requests==2.18.4
tornado==4.5.3
umongo==0.15.0
urllib3==1.22
